+++ packages/core/src/helpers/get-prop-functions.ts
@@ -1,4 +1,4 @@
-import traverse from 'traverse';
+import traverse from 'neotraverse/legacy';
 import { MitosisComponent } from '../types/mitosis-component';
 
 const propsRegex = /props\s*\.\s*([a-zA-Z0-9_\4]+)\(/;