+++ packages/core/src/parsers/jsx/element-parser.ts
@@ -26,12 +26,7 @@ const getForArguments = (params: any[]): ForNode['scope'] => {
  * Parses a JSX element into a MitosisNode.
  */
 export const jsxElementToJson = (
-  node:
-    | babel.types.JSXElement
-    | babel.types.JSXText
-    | babel.types.JSXFragment
-    | babel.types.JSXExpressionContainer
-    | babel.types.JSXSpreadChild,
+  node: babel.types.Expression | babel.types.JSX,
 ): MitosisNode | null => {
   if (types.isJSXText(node)) {
     return createMitosisNode({
@@ -40,83 +35,89 @@ export const jsxElementToJson = (
       },
     });
   }
-  if (types.isJSXExpressionContainer(node)) {
-    if (types.isJSXEmptyExpression(node.expression)) {
-      return null;
-    }
-    // foo.map -> <For each={foo}>...</For>
-    if (
-      types.isCallExpression(node.expression) ||
-      types.isOptionalCallExpression(node.expression)
-    ) {
-      const callback = node.expression.arguments[0];
-      if (types.isArrowFunctionExpression(callback)) {
-        if (types.isIdentifier(callback.params[0])) {
-          const forArguments = getForArguments(callback.params);
-          return createMitosisNode({
-            name: 'For',
-            bindings: {
-              each: createSingleBinding({
-                code: generate(node.expression.callee)
-                  .code // Remove .map or potentially ?.map
-                  .replace(/\??\.map$/, ''),
-              }),
-            },
-            scope: forArguments,
-            children: [jsxElementToJson(callback.body as any)!],
-          });
-        }
-      }
-    }
 
-    // {foo && <div />} -> <Show when={foo}>...</Show>
-    if (types.isLogicalExpression(node.expression)) {
-      if (node.expression.operator === '&&') {
-        return createMitosisNode({
-          name: 'Show',
-          bindings: {
-            when: createSingleBinding({ code: generate(node.expression.left).code! }),
-          },
-          children: [jsxElementToJson(node.expression.right as any)!],
-        });
-      } else {
-        // TODO: good warning system for unsupported operators
-      }
-    }
-
-    // {foo ? <div /> : <span />} -> <Show when={foo} else={<span />}>...</Show>
-    if (types.isConditionalExpression(node.expression)) {
-      return createMitosisNode({
-        name: 'Show',
-        meta: {
-          else: jsxElementToJson(node.expression.alternate as any)!,
-        },
-        bindings: {
-          when: createSingleBinding({ code: generate(node.expression.test).code }),
-        },
-        children: [jsxElementToJson(node.expression.consequent as any)!],
-      });
-    }
-
-    // TODO: support {foo ? bar : baz}
-
-    return createMitosisNode({
-      bindings: {
-        _text: createSingleBinding({ code: generate(node.expression).code }),
-      },
-    });
+  if (types.isJSXEmptyExpression(node)) {
+    return null;
   }
 
-  if (types.isJSXFragment(node)) {
+  if (types.isJSXExpressionContainer(node)) {
+    return jsxElementToJson(node.expression as any);
+  }
+
+  if (types.isCallExpression(node) || types.isOptionalCallExpression(node)) {
+    const callback = node.arguments[0];
+    if (types.isArrowFunctionExpression(callback)) {
+      if (types.isIdentifier(callback.params[0])) {
+        const forArguments = getForArguments(callback.params);
+        return createMitosisNode({
+          name: 'For',
+          bindings: {
+            each: createSingleBinding({
+              code: generate(node.callee)
+                .code // Remove .map or potentially ?.map
+                .replace(/\??\.map$/, ''),
+            }),
+          },
+          scope: forArguments,
+          children: [jsxElementToJson(callback.body as any)!],
+        });
+      }
+    }
+  } else if (types.isLogicalExpression(node)) {
+    // {foo && <div />} -> <Show when={foo}>...</Show>
+    if (node.operator === '&&') {
+      return createMitosisNode({
+        name: 'Show',
+        bindings: {
+          when: createSingleBinding({ code: generate(node.left).code! }),
+        },
+        children: [jsxElementToJson(node.right as any)!],
+      });
+    } else {
+      // TODO: good warning system for unsupported operators
+    }
+  } else if (types.isConditionalExpression(node)) {
+    // {foo ? <div /> : <span />} -> <Show when={foo} else={<span />}>...</Show>
+    const child = jsxElementToJson(node.consequent as any);
+    return createMitosisNode({
+      name: 'Show',
+      meta: {
+        else: jsxElementToJson(node.alternate as any),
+      },
+      bindings: {
+        when: createSingleBinding({ code: generate(node.test).code }),
+      },
+      children: child === null ? [] : [child],
+    });
+  } else if (types.isJSXFragment(node)) {
     return createMitosisNode({
       name: 'Fragment',
       children: node.children.map(jsxElementToJson).filter(checkIsDefined),
     });
-  }
-
-  // TODO: support spread attributes
-  if (types.isJSXSpreadChild(node)) {
+  } else if (types.isJSXSpreadChild(node)) {
+    // TODO: support spread attributes
     return null;
+  } else if (types.isNullLiteral(node) || types.isBooleanLiteral(node)) {
+    return null;
+  } else if (types.isNumericLiteral(node)) {
+    return createMitosisNode({
+      properties: {
+        _text: String(node.value),
+      },
+    });
+  } else if (types.isStringLiteral(node)) {
+    return createMitosisNode({
+      properties: {
+        _text: node.value,
+      },
+    });
+  }
+  if (!types.isJSXElement(node)) {
+    return createMitosisNode({
+      bindings: {
+        _text: createSingleBinding({ code: generate(node).code }),
+      },
+    });
   }
 
   const nodeName = generate(node.openingElement.name).code;
@@ -186,51 +187,63 @@ export const jsxElementToJson = (
     }
   }
 
-  return createMitosisNode({
-    name: nodeName,
-    properties: node.openingElement.attributes.reduce<MitosisNode['properties']>((memo, item) => {
-      if (types.isJSXAttribute(item)) {
-        const key = transformAttributeName(item.name.name as string);
-        const value = item.value;
-        if (types.isStringLiteral(value)) {
-          memo[key] = value.value;
-          return memo;
-        }
-        if (types.isJSXExpressionContainer(value) && types.isStringLiteral(value.expression)) {
-          memo[key] = value.expression.value;
-          return memo;
-        }
-      }
-      return memo;
-    }, {}),
-    bindings: node.openingElement.attributes.reduce<MitosisNode['bindings']>((memo, item) => {
+  // const properties: MitosisNode['properties'] = {}
+  // const bindings: MitosisNode['bindings'] = {}
+  // const slots: MitosisNode['slots'] = {}
+
+  const { bindings, properties, slots } = node.openingElement.attributes.reduce<{
+    bindings: MitosisNode['bindings'];
+    properties: MitosisNode['properties'];
+    slots: {} & MitosisNode['slots'];
+  }>(
+    (memo, item) => {
       if (types.isJSXAttribute(item)) {
         const key = transformAttributeName(item.name.name as string);
         const value = item.value;
 
-        // boolean attribute
+        // <Foo myProp />
         if (value === null) {
-          memo[key] = createSingleBinding({ code: 'true' });
+          memo.bindings[key] = createSingleBinding({ code: 'true' });
           return memo;
         }
-        if (types.isJSXExpressionContainer(value) && !types.isStringLiteral(value.expression)) {
-          const { expression } = value;
-          if (types.isArrowFunctionExpression(expression)) {
-            if (key.startsWith('on')) {
-              const args = expression.params.map((node) => (node as babel.types.Identifier)?.name);
-              memo[key] = createSingleBinding({
-                code: generate(expression.body).code,
-                arguments: args.length ? args : undefined,
-              });
-            } else {
-              memo[key] = createSingleBinding({ code: generate(expression.body).code });
-            }
-          } else {
-            memo[key] = createSingleBinding({ code: generate(expression).code });
-          }
 
+        // <Foo myProp="hello" />
+        if (types.isStringLiteral(value)) {
+          memo.properties[key] = value.value;
           return memo;
         }
+
+        if (!types.isJSXExpressionContainer(value)) return memo;
+
+        const { expression } = value;
+
+        if (types.isStringLiteral(expression)) {
+          // <Foo myProp={"hello"} />
+          memo.properties[key] = expression.value;
+        } else if (types.isArrowFunctionExpression(expression)) {
+          // <Foo myProp={() => {}} />
+          const args = key.startsWith('on')
+            ? expression.params.map((node) => (node as babel.types.Identifier)?.name)
+            : [];
+
+          memo.bindings[key] = createSingleBinding({
+            code: generate(expression.body).code,
+            arguments: args.length ? args : undefined,
+          });
+        } else if (types.isJSXElement(expression)) {
+          // <Foo myProp={<MoreMitosisNode><div /></MoreMitosisNode>} />
+          const slotNode = jsxElementToJson(expression);
+          if (!slotNode) return memo;
+
+          memo.slots[key] = [slotNode];
+
+          // Temporarily keep the slot as a binding until we migrate generators to use the slots.
+          memo.bindings[key] = createSingleBinding({ code: generate(expression).code });
+        } else {
+          memo.bindings[key] = createSingleBinding({ code: generate(expression).code });
+        }
+
+        return memo;
       } else if (types.isJSXSpreadAttribute(item)) {
         // TODO: potentially like Vue store bindings and properties as array of key value pairs
         // too so can do this accurately when order matters. Also tempting to not support spread,
@@ -238,14 +251,26 @@ export const jsxElementToJson = (
 
         const { code: key } = generate(item.argument);
 
-        memo[key] = {
+        memo.bindings[key] = {
           code: types.stringLiteral(generate(item.argument).code).value,
           type: 'spread',
           spreadType: 'normal',
         };
       }
       return memo;
-    }, {}),
+    },
+    {
+      bindings: {},
+      properties: {},
+      slots: {},
+    },
+  );
+
+  return createMitosisNode({
+    name: nodeName,
+    properties,
+    bindings,
     children: node.children.map(jsxElementToJson).filter(checkIsDefined),
+    slots: Object.keys(slots).length > 0 ? slots : undefined,
   });
 };